frontend/api/requirements.txt
azure-functions
pydantic==2.5.0
typing-extensions